* Refactor scripts UI Tweek
All managers are now Singletons.
The HUD canvas is integrated to the MainMenu prefab for better implementation as it is now behind the other panels.
The HUDComponent script is obsolete, component has been integrated to The HUDmanager.
Food and money text are updated Only if there is a change, not every frame.
UpdateMoneyText method added to the money script.
Added a Warm Up Count down text. (need more tweek)
It is time to put aside the PlayScene for now, I made a _1NewPlayScene in my folder (_Alex), copy that one if you need to work on stuff.