* Refactoring
NewSpawnManager now has a OnDestroy to that turns the Singleton instance null, so when a new playground is place on with its spawn manager a new singleton is made from it.
EnemyMovementMechanics some refactoring to take advantage of singleton, also remove remains of testing.
Tile prefab has now a convex mesh collider to prevent error since we are using scale -1 when flipping the model on X or Z during tiling. (box collider do not support negative scaling)