+* New Models, Script change, new Materials. New prefabs. Moved files
I made a folder called Obsolete for the old play scene.
I added 2 new prefabs with model for the toy soldier level 2 and 3 , prefabs are working towers.
Improved the damage over time coroutine it exit if enemy reached 0 hitpoints.
modifed the camera script to use fixedDeltatime so if we scale time the camera still move at the same speed.
Modifed the GameManager so that when you buy an item the store disapears, it gives feedback to player that buying was succssfull.