+* Tile Automatic Tool, Tag, Prefabs, BSG Editor
The new Tile tool is online, There is a menu (TileAutomatic) with refresh tile, setup, and reset tiles. Snap the tiles next to each other and hit refresh and the tool does it's magic. There is 3 diferent tiles. Tower, Path an Bush, the path one is always straight it will not change, and is unafected by reset. bush and Tower will not connected to each other they are independed. The TilePath can be rotated and set to unavailable for barriers allignement purposes, through the used of a check mark on the inspector.
Added TileBush Tags, and set appropriate tags to tile prefabs.
Updated the BSG Editor, removed obsolete functions.