* TileAutomatic Tool Improvement and more
I added Snap Tiles to the menu, it now snaps the tiles to the nearest  point based on the tileSize, This will help fix if some tiles are not lined up properly next to each others.
Fix some missing objects/model on ToySoldier level 2 and 3 prefabs.
Changed newSpawnManager to disable warmup text gameobject instead of the text component, after the countdown.