-+* Added Flyer Ant, Reworked the Projectile Script, and more.
We  now have a flying enemy, it moves faster and fly over barriers. Added its models, materials, animations and prefab, also had to create a new EnemyFlyer script for it.
The projectile.cs for its movement mechanics was creating problem since the start since it needed a target to move when the enemy would die it wouldnt continue on it's path. So at first I made it spawn a new gameobject to continue to there and collide with the ground but then for the flyer ant it would just stay mid air. Now all projective moves with force using their rigidbody, tons of code was replace with 1 line, including in Tower.cs.
Placed walls and ground on the Backyard prefab so the stray projectiles can collide with them. Added a Terrain tag to take care of that.
Fixed bug where the enemy would not rotate randomly when diying,
Increased Cheerios' colliders height to take the flyer and into consideration.
Reworked the Lego Gun Level 3 prefas so it fits properly on a tile.
Reworked the SO for shooter and laser and the projectiles prefabsto have proper speed for new mechanics. (could still need more balances)
Added a gizmo to the Laser GO prefab so we can see in the editor view for debuging purpuses.