-+* Range Shader and Upgrade Methods
You can now upgrade the towers buy simply clicking buy again, proper UI integration need to be done, selectedTile.CurrentItem.GetComponent<Item>().UpgradeVersion points the the upgrade version Item, this way you can retrieve value (price)
The Towers now display their range, as a Capsule with a custom Shader, Color can be change and this shader can be customized to be used for other purposes. Placeholder will show the range and when you select a tower it will show it range, Upgraded towers show their upgraded range.
Changed Money text on HUD to be best fit and overflow, it was on truncate.