+* Change Prefabs, Removed Obsolete code, Fix Code
I have changed the flyer prefab to a sphere colider since its more performant.
Added some to the projectile script some protection as it was possible to generate an error.
Modified the Ant Exit prefab so the branch lines up with the exit so the flying ant goes that way.
Made a test prefab from level 1 and modified it to fix the waypoint outside the navigation mesh.
Tag the GameManager as root to be used as a spawn point for the level.