-+* Added UpgradeItemInfo Script
Now when you highlight/select the buy button there is a script, "UpgradeItemInfo", that detects this  and can do stuff, for now it shows the RangeUpgrade for the next level. and its on the playmenu prefab on the buy button.
Added a RangeUpgrade GameObject to all the Towers prefab.
Added a RangeUpgrade material for it.
Improved the range scaling methods, now a broadcast from the towerscript.
Had to change the prefab from the Magnifyer Placeholder as it was not a variant from the proper prefab. Save that to the GameManager prefab.
Deleted a test script.
Moved SimpleDisable.cs to Obsolete.