+* Added enemies Health Bars and more
Enemies now have an health bar and the EnemyBaseClass is taking care of it. The GameManager has a showHealthBar bool and a togglehealthbar method to show or hide them. All enemy prefabs have been modified to suit this change. There is a Health bar Prefab. The enemybaseclass has a starting and ending color for health bar that can be modified as needed
Greatly improved the UpgradeItemInfo script, now using Ipointer interface enter and exit.
Some misc change on my folder.