* Fix projectile bug and worker collider missplace
Alright an other itteration of the projectile script, this one should be the final one, It never misses for bullet and laser, and the missile speed is properly working.
The Ant Worker collider now touches the floor so bullet do not go under and miss it when comming from far away.