+* New Queen Enemy and more
We now have a Queen that lays egg and they spawn enemies from it. New Model, Prefab, Material, Animation and script for it.
I have adjusted all Enemies prefab NavMeshAgent settings, removed the overwrite from the script. All the enemies are now Agent Type: Ant. The Nav mesh need to be rebaked with this in mind,
The Ant Entrance model had to be modified, the Z axis (forward) was not orriented properly ants were spawning sideways.
The main menu scene has been fix since I modifed the Ant Model with a Health Bar.
GameManager prefab is no longer tag Root, since the Backyard has a root point.
Fixed a Bug on the prefab Level 2, one of the Lego piece was not attached on the barrel.
Renamed the Magnifying Glass Laser.