* Fixes
Minor fix to the Range scaling method, properly automaticaly scale to the set range value, clamped the Z size to 10 unit to improve look.
Improved performance on the method that check if the level is completed in the enemy manager script
Removed some old commented code in GameManager.
Added a property of the current level to the levelmanager so that Queen's eggs are now spawned as children of the level so when a new level is loaded the eggs are also destroyed.
Removed a redundant method in the MoneyManager and improved performance.