* Bug fixed
Added to the item script an enum to tell if the ITEM is normal or a placeholder, this is so the range GO gets the proper scaling and to optimize the script, since the component gets destroyed for the placeholder as they are no longer needed. Updated the prefabs to reflect the changes.
Added a header to the playerbase class the help visual on inspector.
Fix a bug on the newspawnmanager where it would not tell all the enemy were spawned right after they where, it waited a spawning rate circle before,
Added level completed reset to the levelmanager as it spawn a new level.