*Fix Spawning problem
OK not this is properly fix and work as intented, the enemy spawn right away at the begining of the wave and after the last enemy it waits as per the rate to start the next wave. but if the enemy is the last of the level it says that the spawning is completed.
This might change since Leo is reworking the spawner