* Simple fixes.
Some small fix to Level prefabs, Entrance, Exit and Cheerios were unpacked (no longer prefabs) some path tiles where missaligned and some corner where "available" for barriers, Extended the exit so ants disapeared once they enter the LOG and not before. 
My last fix to the spawn manager revised with Leo and properly implemented.
Cheerios are now ignoring raycast as it should.
Repositioned the Root of the Backyard level to be more in the center of the patch without grass