+* New hud buttons fixes and commenting.
We now have a HP toggle button, shows or hide the enemies health bars.
We now have a fast forward button that increase the speed as per set on the GameManager, And the script is ready to receive the function to skip the Level warmup sequence. Menu System Playscene prefab updated.
Minor fixes on the level prefabs.
When you hit esc it now either deselect a tile or pause and unpause appropriately.
Added 'p' as a key input for "Cancel", 'P' for pause.