* Code Cleanup, fixes and more
Change the Crunch particle to Local from World space, now it moves with the ant.
Change the misile sound to be part of the VFX GameObject so it is now 3D sound and part of the VFX sound track.
Rename Pause.cs to PauseManager and it should be.
Fixed the Mouse sensitivity setting that was not being used aproperly. and change the setting to default to 25 sensitivity and the slider goes from 10 to 50
Cleaned the code, made public variable private, serilized and public properties. removed serialized from some of properties that did not need to be.
Added descriptions to all the methods on most scripts.
Removed old commented obsolete code lines.